requirements.txt
# python module requiments
discord.py
jsons
requests
